Refactoring for Cppcheck's graphical user interface
Kimmo Varis contributed the following details for the discussion topic
“GUI: Ideas to improve?” on 2011-07-17.
http://cppcheck.sourceforge.net/forum/viewtopic.php?f=3&t=237&p=1432&sid=f45436dbecc246342f6a423d12d2226e#p1432

“…
As I mentioned in above post there is one major drawback in current GUI code.
The current code does not follow Qt:s Model-View architecture. In practice
we store the error data directly to the View and just use default standard model.
But it should be other way around - data is stored to Model and we should use
the default View code. This severely limits what we can do (efficiently)
in GUI currently. This must be fixed, no way around it.
…”


I got into the mood once more to adjust this software situation a bit.
The source code structure was extended so that the member variable “mModel”
could be deleted from the class “ResultsTree” finally.
The shown implementation contains still some open issues for further
development considerations.

See also:
* Commit b39c15410bb853d69f3d44953e6eaf6220aed491 (by Vesa Pikki on 2009-03-01)
  Began implementing a simple Qt based GUI.

* Extensions around data models and views
  https://trac.cppcheck.net/ticket/8533

Signed-off-by: Markus Elfring <elfring@users.sourceforge.net>